Update connector for Ballerina 1.2.0
