+++ cassandra/Ballerina.lock
@@ -1,0 +1,4 @@
+org_name = "ballerina"
+version = "0.8.5"
+lockfile_version = "1.0.0"
+ballerina_version = "1.2.0"
