+++ cassandra/Ballerina.toml
@@ -1,0 +1,13 @@
+[project]
+org-name="ballerina"
+version="0.8.5"
+
+[platform]
+target = "java8"
+
+[[platform.libraries]]
+module = "cassandra"
+path = "../cassandra-utils/target/cassandra-utils-0.8.5.jar"
+artafactId = "cassandra-utils"
+version = "0.8.5"
+groupId = "org.ballerinalang"
