--- component/src/main/ballerina/Ballerina.toml
@@ -1,3 +1,0 @@
-[project]
-org-name="wso2"
-version="0.0.0"
