--- component/src/main/resources/installation-scripts/README.txt
@@ -1,12 +1,0 @@
-Instructions for usage
-
-This guide contains instructions to install the Cassandra Ballerina module into your Ballerina installation. This is a one-time task.
-Once done, you can then use this module across all your Ballerina projects.
-
-In this directory, you can find two scripts of interest to you. Please note that when running the following scripts,
-you will need write permission to the Ballerina installation location.
-
-1. install.{sh/bat} - Installs the Cassandra module.
-2. uninstall.{sh/bat} - Uninstalls the Cassandra module if installed already.
-
-Happy coding!
